fix(cdk-experimental/ui-patterns): fix lint and build config
