+++ src/cdk-experimental/deferred-content/deferred-content.ts
@@ -7,7 +7,6 @@
  */
 
 import {
-  computed,
   Directive,
   effect,
   inject,
