+++ src/cdk-experimental/ui-patterns/public-api.ts
@@ -9,6 +9,4 @@
 export * from './listbox/listbox';
 export * from './listbox/option';
 export * from './behaviors/signal-like/signal-like';
-export * from './tabs/tab';
-export * from './tabs/tablist';
-export * from './tabs/tabpanel';
+export * from './tabs/tabs';
