+++ src/components-examples/cdk-experimental/tabs/cdk-tabs/cdk-tabs-example.css
